feat: section과 line 분리
